+++ packages/backend/src/nest/storage/channels/messages/public-channel-messages.service.ts
@@ -66,7 +66,7 @@ export class PublicChannelMessagesService extends BaseMessagesService {
       const chain = this.sigChainService.getActiveChain()
       const encryptable: EncryptableMessageComponents = {
         id: rawMessage.id,
-        userId: chain.user.userId,
+        userId: rawMessage.userId,
         type: rawMessage.type,
         channelId: rawMessage.channelId,
         message: rawMessage.message,