+++ CHANGELOG.md
@@ -6,18 +6,24 @@
 
 * Allow all users to create private channels [#3277](https://github.com/TryQuiet/quiet/issues/3277)
 * Allow channel members to add members to private channels [#3300](https://github.com/TryQuiet/quiet/issues/3300)
-* Require team ID on invite links, use team ID for all chain operations, and hash team name on sigchains [#3296](https://github.com/TryQuiet/quiet/issues/3296)
 
 ### Fixes
 
 * Don't send deletion message for private channels [#3273](https://github.com/TryQuiet/quiet/issues/3273)
 * Ensure notification registration waits for auth handshake [#3289](https://github.com/TryQuiet/quiet/issues/3289)
 * Fixed a race condition that can cause stale data to remain after leaving community [#3253](https://github.com/TryQuiet/quiet/issues/3253)
+* Validate user ID on decrypted message matches the signature [#3334](https://github.com/TryQuiet/quiet/issues/3334)
 
 ### Chores
 
 * Enable QSS on prod
 * Add script for cleaning compiled/generated code directories
+
+### Breaking
+
+* Include team ID and createdAt in message encryption and validate on consume [#3304](https://github.com/TryQuiet/quiet/issues/3304)
+* Require team ID on invite links, use team ID for all chain operations, and hash team name on sigchains [#3296](https://github.com/TryQuiet/quiet/issues/3296)
+* Use randomly generated Base58 usernames on sigchain [#3321](https://github.com/TryQuiet/quiet/issues/3321)
 
 ## [7.3.0]